+++ joss/paper.bib
@@ -380,15 +380,15 @@
   language  = {english},
 }
 
-@inproceedings{Sheshadri:13,
+@article{Sheshadri:13,
   author    = {Sheshadri, Aashish and Lease, Matthew},
   title     = {{SQUARE: A Benchmark for Research on Computing Crowd Consensus}},
   year      = {2013},
-  booktitle = {First AAAI Conference on Human Computation and Crowdsourcing},
-  series    = {HCOMP~2013},
+  journal   = {Proceedings of the AAAI Conference on Human Computation and Crowdsourcing},
+  volume    = {1},
+  number    = {1},
   pages     = {156--164},
-  publisher = {Association for the Advancement of Artificial Intelligence},
-  url       = {https://ojs.aaai.org/index.php/HCOMP/article/view/13088},
+  doi       = {10.1609/hcomp.v1i1.13088},
   language  = {english},
 }
 
