+++ tests/postprocessing/test_entropy_threshold.py
@@ -1,65 +1,68 @@
 import pandas as pd
 import pytest
 
+from crowdkit.aggregation import MajorityVote
 from crowdkit.metrics.data import consistency
 from crowdkit.postprocessing import entropy_threshold
-from crowdkit.aggregation import MajorityVote
 
 
 class TestEntropyThreshold:
     def test_entropy_threshold_docstring_test(self) -> None:
         answers = pd.DataFrame.from_records(
             [
-                {'task': '1', 'worker': 'A', 'label': frozenset(['dog'])},
-                {'task': '1', 'worker': 'B', 'label': frozenset(['cat'])},
-                {'task': '2', 'worker': 'A', 'label': frozenset(['cat'])},
-                {'task': '2', 'worker': 'B', 'label': frozenset(['cat'])},
-                {'task': '3', 'worker': 'A', 'label': frozenset(['dog'])},
-                {'task': '3', 'worker': 'B', 'label': frozenset(['cat'])},
+                {"task": "1", "worker": "A", "label": frozenset(["dog"])},
+                {"task": "1", "worker": "B", "label": frozenset(["cat"])},
+                {"task": "2", "worker": "A", "label": frozenset(["cat"])},
+                {"task": "2", "worker": "B", "label": frozenset(["cat"])},
+                {"task": "3", "worker": "A", "label": frozenset(["dog"])},
+                {"task": "3", "worker": "B", "label": frozenset(["cat"])},
             ]
         )
-        with pytest.warns(UserWarning, match='Removed >= 1/2 of answers with entropy_threshold. This might lead to poor annotation quality. '):
+        with pytest.warns(
+            UserWarning,
+            match="Removed >= 1/2 of answers with entropy_threshold. This might lead to poor annotation quality. ",
+        ):
             filtered_answers = entropy_threshold(answers)
-        assert filtered_answers.columns.tolist() == ['task', 'worker', 'label']
+        assert filtered_answers.columns.tolist() == ["task", "worker", "label"]
         assert filtered_answers.shape == (3, 3)
-        assert 'B' not in filtered_answers.worker
+        assert "B" not in filtered_answers.worker
 
     def test_entropy_threshold_consistency_improves(self) -> None:
         answers = pd.DataFrame.from_records(
             [
-                {'task': '1', 'worker': 'A', 'label': frozenset(['dog'])},
-                {'task': '1', 'worker': 'B', 'label': frozenset(['cat'])},
-                {'task': '1', 'worker': 'C', 'label': frozenset(['cat'])},
-                {'task': '1', 'worker': 'D', 'label': frozenset(['dog'])},
-
-                {'task': '2', 'worker': 'A', 'label': frozenset(['cat'])},
-                {'task': '2', 'worker': 'B', 'label': frozenset(['cat'])},
-                {'task': '2', 'worker': 'C', 'label': frozenset(['cat'])},
-                {'task': '2', 'worker': 'D', 'label': frozenset(['dog'])},
-
-                {'task': '3', 'worker': 'A', 'label': frozenset(['dog'])},
-                {'task': '3', 'worker': 'B', 'label': frozenset(['cat'])},
-                {'task': '3', 'worker': 'C', 'label': frozenset(['dog'])},
-                {'task': '3', 'worker': 'D', 'label': frozenset(['dog'])},
-
-                {'task': '4', 'worker': 'A', 'label': frozenset(['dog'])},
-                {'task': '4', 'worker': 'B', 'label': frozenset(['cat'])},
-                {'task': '4', 'worker': 'C', 'label': frozenset(['dog'])},
-                {'task': '4', 'worker': 'D', 'label': frozenset(['dog'])},
+                {"task": "1", "worker": "A", "label": frozenset(["dog"])},
[…]
