+++ crowdkit/aggregation/texts/rover.py
@@ -1,10 +1,10 @@
 __all__ = [
-    'ROVER',
+    "ROVER",
 ]
 
 from copy import deepcopy
 from enum import Enum, unique
-from typing import List, Callable, Dict, Optional, Tuple, cast
+from typing import Callable, Dict, List, Optional, Tuple, cast
 
 import attr
 import numpy as np
@@ -16,10 +16,10 @@ from ..base import BaseTextsAggregator
 
 @unique
 class AlignmentAction(Enum):
-    DELETION = 'DELETION'
-    SUBSTITUTION = 'SUBSTITUTION'
-    INSERTION = 'INSERTION'
-    CORRECT = 'CORRECT'
+    DELETION = "DELETION"
+    SUBSTITUTION = "SUBSTITUTION"
+    INSERTION = "INSERTION"
+    CORRECT = "CORRECT"
 
 
 @attr.s
@@ -69,7 +69,7 @@ class ROVER(BaseTextsAggregator):
     # Available after fit
     # texts_
 
-    def fit(self, data: pd.DataFrame) -> 'ROVER':
+    def fit(self, data: pd.DataFrame) -> "ROVER":
         """Fits the model. The aggregated results are saved to the `texts_` attribute.
 
         Args:
@@ -81,19 +81,21 @@ class ROVER(BaseTextsAggregator):
         """
 
         result = {}
-        grouped_tasks = data.groupby('task') if self.silent else tqdm(data.groupby('task'))
+        grouped_tasks = (
+            data.groupby("task") if self.silent else tqdm(data.groupby("task"))
+        )
         for task, df in grouped_tasks:
-            hypotheses = [self.tokenizer(text) for i, text in enumerate(df['text'])]
+            hypotheses = [self.tokenizer(text) for i, text in enumerate(df["text"])]
 
             edges = self._build_word_transition_network(hypotheses)
             rover_result = self._get_result(edges)
 
-            text = self.detokenizer([value for value in rover_result if value != ''])
+            text = self.detokenizer([value for value in rover_result if value != ""])
 
             result[task] = text
 
-        texts = pd.Series(result, name='text')
-        texts.index.name = 'task'
+        texts = pd.Series(result, name="text")
+        texts.index.name = "task"
         self.texts_ = texts
 
         return self
@@ -114,8 +116,12 @@ class ROVER(BaseTextsAggregator):
         self.fit(data)
         return self.texts_
 
-    def _build_word_transition_network(self, hypotheses: List[List[str]]) -> List[Dict[str, AlignmentEdge]]:
-        edges = [{edge.value: edge} for edge in self._get_edges_for_words(hypotheses[0])]
+    def _build_word_transition_network(
+        self, hypotheses: List[List[str]]
+    ) -> List[Dict[str, AlignmentEdge]]:
+        edges = [
+            {edge.value: edge} for edge in self._get_edges_for_words(hypotheses[0])
+        ]
 
         for sources_count, hyp in enumerate(hypotheses[1:], start=1):
             edges = self._align(edges, self._get_edges_for_words(hyp), sources_count)
@@ -128,9 +134,9 @@ class ROVER(BaseTextsAggregator):
 
     @staticmethod
     def _align(
-            ref_edges_sets: List[Dict[str, AlignmentEdge]],
-            hyp_edges: List[AlignmentEdge],
-            sources_count: int
+        ref_edges_sets: List[Dict[str, AlignmentEdge]],
+        hyp_edges: List[AlignmentEdge],
+        sources_count: int,
     ) -> List[Dict[str, AlignmentEdge]]:
         """Sequence alignment algorithm implementation.
 
[…]
