### 2023-12-09 #5898 @arfon
@stsievert @chrislintott @hoechenberger @stain @jorgedch –
:wave: would any of you be willing to review this submission for JOSS? The submission under consideration is Learning from Crowds with Crowd-Kit (https://github.com/openjournals/joss-reviews/issues/5898)

Many thanks
Arfon
