+++ crowdkit/aggregation/classification/glad.py
@@ -100,21 +100,25 @@ class GLAD(BaseClassificationAggregator):
     n_iter: int = attr.ib(default=100)
     tol: float = attr.ib(default=1e-5)
     silent: bool = attr.ib(default=True)
-    labels_priors: Optional['pd.Series[Any]'] = attr.ib(default=None)
-    alphas_priors_mean: Optional['pd.Series[Any]'] = attr.ib(default=None)
-    betas_priors_mean: Optional['pd.Series[Any]'] = attr.ib(default=None)
+    labels_priors: Optional["pd.Series[Any]"] = attr.ib(default=None)
+    alphas_priors_mean: Optional["pd.Series[Any]"] = attr.ib(default=None)
+    betas_priors_mean: Optional["pd.Series[Any]"] = attr.ib(default=None)
     m_step_max_iter: int = attr.ib(default=25)
     m_step_tol: float = attr.ib(default=1e-2)
 
     # Available after fit
     # labels_
     probas_: Optional[pd.DataFrame] = attr.ib(init=False)
-    alphas_: 'pd.Series[Any]' = named_series_attrib(name="alpha")
-    betas_: 'pd.Series[Any]' = named_series_attrib(name="beta")
+    alphas_: "pd.Series[Any]" = named_series_attrib(name="alpha")
+    betas_: "pd.Series[Any]" = named_series_attrib(name="beta")
     loss_history_: List[float] = attr.ib(init=False)
 
     def _join_all(
-        self, data: pd.DataFrame, alphas: 'pd.Series[Any]', betas: 'pd.Series[Any]', priors: 'pd.Series[Any]'
+        self,
+        data: pd.DataFrame,
+        alphas: "pd.Series[Any]",
+        betas: "pd.Series[Any]",
+        priors: "pd.Series[Any]",
     ) -> pd.DataFrame:
         """Makes a data frame with format `(task, worker, label, variable) -> (alpha, beta, posterior, delta)`"""
         labels = list(priors.index)
@@ -152,7 +156,7 @@ class GLAD(BaseClassificationAggregator):
         # sum up by workers
         probas = data.groupby(["task", "variable"]).sum(numeric_only=True)["posterior"]
         # add priors to every label
-        probas = probas.add(np.log(cast('pd.Series[Any]', self.priors_)), level=1)
+        probas = probas.add(np.log(cast("pd.Series[Any]", self.priors_)), level=1)
         # exponentiate and normalize
         probas = probas.groupby(["task"]).transform(self._softmax)
         # put posterior in data['posterior']
@@ -166,7 +170,7 @@ class GLAD(BaseClassificationAggregator):
 
     def _gradient_Q(
         self, data: pd.DataFrame
-    ) -> Tuple['pd.Series[Any]', 'pd.Series[Any]']:
+    ) -> Tuple["pd.Series[Any]", "pd.Series[Any]"]:
         """Computes gradient of loss function"""
 
         sigma = scipy.special.expit(data["alpha"] * np.exp(data["beta"]))
@@ -180,15 +184,15 @@ class GLAD(BaseClassificationAggregator):
         dQbeta = data.groupby("task").sum(numeric_only=True)["dQb"]
 
         # gradient of priors on betas
-        assert self.betas_ is not None, 'betas_ is None'
-        assert self.betas_priors_mean_ is not None, 'betas_priors_mean_ is None'
+        assert self.betas_ is not None, "betas_ is None"
+        assert self.betas_priors_mean_ is not None, "betas_priors_mean_ is None"
         dQbeta -= self.betas_ - self.betas_priors_mean_
 
         data["dQa"] = data["posterior"] * (data["delta"] - sigma) * np.exp(data["beta"])
[…]
