+++ .github/workflows/release.yml
@@ -8,7 +8,7 @@ jobs:
   release:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
@@ -28,7 +28,7 @@ jobs:
   docs:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
@@ -39,7 +39,7 @@ jobs:
           pip install pipenv
           pipenv install --dev --system --skip-lock
       - name: Checkout Toloka/docs
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
         with:
           repository: "Toloka/docs"
           path: "docs-repo"
