+++ .github/workflows/tests.yml
@@ -45,6 +45,7 @@ jobs:
         with:
           name: crowd-kit-jupyter
           path: examples/*.html
+        if: matrix.python-version == '3.8'
       - name: Validate CITATION.cff
         uses: citation-file-format/cffconvert-github-action@2.0.0
         with:
