### 2023-12-16 07c4240
Move .isort.cfg to setup.cfg
