+++ .isort.cfg
@@ -1,3 +1,0 @@
-[settings]
-profile=black
-src_paths=crowdkit,tests
