+++ setup.cfg
@@ -80,3 +80,7 @@ strict = True
 select = E,W,F
 ignore = E122,E123,E127,E131,E203,E225,E226,E24,E275,E305,E306,E402,E722,E731,E741,F722,W503,W504,C9,N8
 max-line-length = 200
+
+[isort]
+profile=black
+src_paths=crowdkit,tests
