### 2024-01-08 #5898 @dustalov
HI @arfon, thank you for letting me know. I checked the [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) and have three questions:

1. if the potential reviewer has no business relationship with us, but they are a user of our library, would it be a COI? (I guess not, but let's clarify this, please)
2. is it sufficient for the reviewers to write to this issue?
3. how many reviewers do we need?
