### 2024-01-09 #5898 @arfon
> if the potential reviewer has no business relationship with us, but they are a user of our library, would it be a COI? (I guess not, but let's clarify this, please)

Users are absolutely fine – in fact this can make for a great reviewer.

> is it sufficient for the reviewers to write to this issue?

Reviews happen on GitHub yes. If you can suggest GitHub handles here I can ping them to ask.

> how many reviewers do we need?

At least two, and ideally three.
