### 2024-01-13 #6227 @editorialbot
[REVIEW]: Learning from Crowds with Crowd-Kit
**Submitting author:** @dustalov (<a href="http://orcid.org/0000-0002-9979-2188">Dmitry Ustalov</a>)
**Repository:** https://github.com/Toloka/crowd-kit
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.1
