### 2024-01-13 #6227 @arfon
@jorgedch – This is the review thread for the paper. All of our communications will happen here from now on.

Also, noting here that @jorgedch believes it will take them more like 6-8 weeks to complete their review.
