### 2024-01-31 #6227 @arfon
@neothethird @Mind-the-Cap - :wave: would you be willing to review this submission for JOSS? The submission under consideration is _Learning from Crowds with Crowd-Kit_

Many thanks
Arfon
