### 2024-02-02 #6227 @Mind-the-Cap
Hi @dustalov, I see Boris Tseitlin in the list of the authors but I don't see his account in the contributors list https://github.com/Toloka/crowd-kit/graphs/contributors

I suspect some code has been copied from an internal repo, would you be able to confirm that or to give me the explanation? Many thanks!
