### 2024-02-02 Toloka/crowd-kit#98 @Mind-the-Cap
[DOCS] test datasets
### Problem description

Hi, I'm reviewing the library for JOSS https://github.com/openjournals/joss-reviews/issues/6227

In your tutorials, you suggest using labeled_train_data.tsv, however I cannot find this data. Is is provided somewhere?

Thanks!

### Documentation links

https://github.com/Toloka/crowd-kit/blob/main/examples/ECIR2023-Intents.ipynb

### Potential fix suggestion

_No response_
