### 2024-02-03 Toloka/crowd-kit#98 @dustalov
@denaxen @aliskin @pilot7747 I could not find the ECIR '23 tutorial dataset at https://github.com/clinc/oos-eval/tree/master/data and https://toloka.ai/events/ecir-tutorial-2023/. Maybe we have these files, `labeled_train_data.tsv` and `labeled_test_data.tsv `, anywhere in Toloka's Google Drive or S3?
