+++ setup.cfg
@@ -1,86 +1,0 @@
-[metadata]
-name = crowd-kit
-version = attr: crowdkit.__version__
-description = Computational Quality Control for Crowdsourcing
-long_description = file: README.md
-long_description_content_type = text/markdown
-keywords = crowdsourcing, quality control, answer aggregation
-license = Apache 2.0
-url = https://github.com/Toloka/crowd-kit
-author = Toloka
-classifiers =
-    Development Status :: 5 - Production/Stable
-    Intended Audience :: Developers
-    Intended Audience :: Science/Research
-    License :: OSI Approved :: Apache Software License
-    Operating System :: OS Independent
-    Programming Language :: Python
-    Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.8
-    Programming Language :: Python :: 3.9
-    Programming Language :: Python :: 3.10
-    Programming Language :: Python :: 3.11
-    Programming Language :: Python :: Implementation :: CPython
-    Topic :: Scientific/Engineering
-    Topic :: Software Development
-    Typing :: Typed
-project_urls =
-    Bug Tracker = https://github.com/Toloka/crowd-kit/issues
-    Documentation = https://toloka.ai/en/docs/crowd-kit/
-    API Reference = https://toloka.ai/en/docs/crowd-kit/
-    Source Code = https://github.com/Toloka/crowd-kit
-    Release Notes = https://github.com/Toloka/crowd-kit/blob/main/CHANGELOG.md
-
-[options]
-zip_safe = True
-install_requires =
-    attrs
-    numpy
-    pandas >= 1.1.0
-    tqdm
-    scikit-learn
-    nltk
-    transformers
-python_requires = >= 3.8
-packages = find:
-include_package_data = True
-
-[options.packages.find]
-include = crowdkit*
-
-[options.package_data]
-* = py.typed
-
-[options.extras_require]
-dev =
-    mypy
-    pytest
-    twine
-    flake8
-    black
-    isort
-    pre-commit
-    build
-    codecov
-    notebook
-    ipywidgets
-    stubmaker @ git+https://github.com/Toloka/stubmaker.git@main
-    pandas-stubs
-
-learning =
-    torch>=1.6.0
-    transformers>=4.0.0
-
-[mypy]
-ignore_missing_imports = True
-plugins = numpy.typing.mypy_plugin
-strict = True
-
-[flake8]
-select = E,W,F
-ignore = E122,E123,E127,E131,E203,E225,E226,E24,E275,E305,E306,E402,E722,E731,E741,F722,W503,W504,C9,N8
-max-line-length = 200
-
-[isort]
-profile=black
-src_paths=crowdkit,tests
