+++ tests/learning/test_crowdlayer.py
@@ -13,6 +13,7 @@ def test_crowdlayer_vb(
     assert out.shape == (3, 5)
     assert torch.allclose(out, toy_logits)
     out.sum().backward()
+    assert crowd_layer.weight.grad
     assert crowd_layer.weight.grad.shape == (2, 5)
 
 
@@ -24,6 +25,7 @@ def test_crowdlayer_vw(
     assert out.shape == (3, 5)
     assert torch.allclose(out, toy_logits)
     out.sum().backward()
+    assert crowd_layer.weight.grad
     assert crowd_layer.weight.grad.shape == (2, 5)
 
 
@@ -35,7 +37,9 @@ def test_crowdlayer_vw_b(
     assert out.shape == (3, 5)
     assert torch.allclose(out, toy_logits)
     out.sum().backward()
+    assert crowd_layer.scale.grad
     assert crowd_layer.scale.grad.shape == (2, 5)
+    assert crowd_layer.bias.grad
     assert crowd_layer.bias.grad.shape == (2, 5)
 
 
@@ -47,4 +51,5 @@ def test_crowdlayer_mw(
     assert out.shape == (3, 5)
     assert torch.allclose(out, toy_logits)
     out.sum().backward()
+    assert crowd_layer.weight.grad
     assert crowd_layer.weight.grad.shape == (2, 5, 5)
