+++ crowdkit/aggregation/pairwise/__init__.py
@@ -1,3 +1,7 @@
+"""
+Routines for pairwise aggregation.
+"""
+
 from .bradley_terry import BradleyTerry
 from .noisy_bt import NoisyBradleyTerry
 
