+++ crowdkit/aggregation/multilabel/__init__.py
@@ -1,3 +1,7 @@
+"""
+Routines for multi-label aggregation.
+"""
+
 from .binary_relevance import BinaryRelevance
 
 __all__ = ["BinaryRelevance"]
