+++ crowdkit/learning/__init__.py
@@ -1,3 +1,7 @@
+"""
+Routines for deep learning for crowds.
+"""
+
 from .conal import CoNAL
 from .crowd_layer import CrowdLayer
 from .text_summarization import TextSummarization
