+++ crowdkit/aggregation/image_segmentation/__init__.py
@@ -1,3 +1,7 @@
+"""
+Routines for image segmentation.
+"""
+
 from .segmentation_em import SegmentationEM
 from .segmentation_majority_vote import SegmentationMajorityVote
 from .segmentation_rasa import SegmentationRASA
