+++ misc/described_objects_crowdkit.py
@@ -1,3 +1,0 @@
-from pandas import DataFrame, Series
-
-DESCRIBED_OBJECTS = {DataFrame: ("pandas", "DataFrame"), Series: ("pandas", "Series")}
