+++ crowdkit/aggregation/texts/__init__.py
@@ -1,3 +1,7 @@
+"""
+Routines for textual data.
+"""
+
 from .rover import ROVER
 from .text_hrrasa import TextHRRASA
 from .text_rasa import TextRASA
