+++ crowdkit/aggregation/classification/__init__.py
@@ -1,3 +1,7 @@
+"""
+Routines for classification.
+"""
+
 __all__ = [
     "DawidSkene",
     "GLAD",
