+++ crowdkit/metrics/__init__.py
@@ -1,3 +1,7 @@
+"""
+Routines for metrics.
+"""
+
 from typing import List
 
 __all__: List[str] = []
