+++ .github/workflows/tests.yml
@@ -16,7 +16,6 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
-          cache: pipenv
       - name: Install dependencies
         run: |
           python3 -m pip install --upgrade pip
