### 2024-02-06 #6227 @dustalov
Hi @Mind-the-Cap, Boris and Nikita worked on `crowd-kit[learning]` subpackage together in an internal repo at Toloka (Yandex Toloka at the time). Then, the corresponding code was open-sourced by Nikita.
