### 2024-02-06 Toloka/crowd-kit#98 @dustalov
We found the files in the supplementary materials for our tutorial: <https://drive.google.com/drive/folders/1jMNkCs1DzJESiL2-8Pr8NJhqyi7GZJp5?usp=sharing> (`Labeled data` subdirectory). I can adjust the notebook accordingly.
