### 2024-02-06 Toloka/crowd-kit#98 @dustalov
@Mind-the-Cap fixed in e8086abfd10379db7cafe0276a6e733fbf7c417c, now the files are downloaded using `gdown`.
