### 2024-02-07 #5898 @arfon
@mitchellg  – can you affirm over in https://github.com/openjournals/joss-reviews/issues/6227 that you're happy to be a reviewer here?
