+++ .readthedocs.yaml
@@ -1,0 +1,16 @@
+version: 2
+formats:
+  - epub
+  - pdf
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "3.11"
+mkdocs:
+  configuration: mkdocs.yml
+python:
+  install:
+    - method: pip
+      path: .
+      extra_requirements:
+        - docs
