### 2024-02-11 241c57d
Use nbQA with pre-commit
