+++ .github/workflows/release.yml
@@ -12,12 +12,12 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: "3.8"
+          python-version: "3.11"
       - name: Install dependencies
         run: |
           python3 -m pip install --upgrade pip
           pip install pipenv
-          pipenv install --python "3.8" --dev --system --skip-lock
+          pipenv install --python "3.11" --dev --system --skip-lock
       - name: Build and publish
         env:
           TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
