+++ .pre-commit-config.yaml
@@ -18,3 +18,12 @@ repos:
   rev: 6.0.0
   hooks:
   - id: flake8
+- repo: https://github.com/nbQA-dev/nbQA
+  rev: 1.7.1
+  hooks:
+   - id: nbqa-black
+     additional_dependencies: [black==23.3.0]
+   - id: nbqa-isort
+     additional_dependencies: [isort==5.12.0]
+   - id: nbqa-flake8
+     additional_dependencies: [flake8==6.0.0]
