### 2024-02-11 c15fa61
Use pyupgrade
