+++ .pre-commit-config.yaml
@@ -23,6 +23,8 @@ repos:
   hooks:
    - id: nbqa-black
      additional_dependencies: [black==23.3.0]
+   - id: nbqa-pyupgrade
+     additional_dependencies: [pyupgrade==3.15.0]
    - id: nbqa-isort
      additional_dependencies: [isort==5.12.0]
    - id: nbqa-flake8
