+++ tests/aggregation/test_mace_aggregation.py
@@ -1,6 +1,7 @@
 """
 Simple aggregation tests.
 """
+
 from typing import Any
 
 import numpy as np
