+++ .pre-commit-config.yaml
@@ -22,10 +22,10 @@ repos:
   rev: 1.7.1
   hooks:
    - id: nbqa-black
-     additional_dependencies: [black==23.3.0]
+     additional_dependencies: [black==24.2.0]
    - id: nbqa-pyupgrade
      additional_dependencies: [pyupgrade==3.15.0]
    - id: nbqa-isort
-     additional_dependencies: [isort==5.12.0]
+     additional_dependencies: [isort==5.13.2]
    - id: nbqa-flake8
-     additional_dependencies: [flake8==6.0.0]
+     additional_dependencies: [flake8==7.0.0]
