+++ .github/workflows/release.yml
@@ -12,16 +12,16 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: "3.11"
+          python-version: "3.12"
       - name: Install dependencies
         run: |
           python3 -m pip install --upgrade pip
           pip install pipenv
-          pipenv install --python "3.11" --dev --system --skip-lock
-      - name: Build and publish
-        env:
-          TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
-          TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
-        run: |
-          python3 -m build --sdist --wheel .
-          twine upload dist/*
+          pipenv install --python "3.12" --dev --system --skip-lock
+      - name: Build
+        run: python3 -m build --sdist --wheel .
+      - name: Publish
+        uses: pypa/gh-action-pypi-publish@release/v1
+        with:
+          user: ${{ secrets.PYPI_USERNAME }}
+          password: ${{ secrets.PYPI_PASSWORD }}
