### 2024-02-21 #6227 @mitchellg
This is mostly looking good! The one area in the checklist that I couldn't find was "clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support".
