### 2024-02-22 #6227 @dustalov
Hi @mitchellg, thank you!

1\. We have a `CONTRIBUTING.md` document inside the repository, but `README.md` does not mention it (however, it is present in issue/PR templates). We can state and link it explicitly.

2 & 3\. We accept feature requests and issue reports and offer help via GitHub Issues. I'll make an update to the documentation.

Would these steps address your comments?
