### 2024-03-04 #6227 @jorgedch
Hi @dustalov @arfon,

I've gone through each element in the checklist now. I've ran the unit tests, code coverage (95%), successfully ran the jupyter notebook with the complete tutorial, went through all the documentation and the paper. The text is clear and well referenced.

The project is very mature. All code is well documented with type definitions, docstrings and references where relevant. The installation process runs smoothly as indicated in the README and the process for contributors is now clearly stated after the initial feedback has been addressed. Finally, the aggregation (`GLAD`, `DawidSkene`, `MajorityVote`) and learning (`CoNAL`, `CrowdLayer`) methods ran pretty fast on my local machine.

It would be nice to use a dataset where the f1 score of Dawid-Skene or GLAD have a larger improvement over Majority Vote to understand when is that such methods shine over the baseline. But this is just a nice-to-have, nothing that takes away from the package itself.

I only would add that I'm glad to have found this project and looking forward to integrating it in future developments. Congratulations to all the authors for the great work!

Good to go from my side 👍
