+++ crowdkit/__init__.py
@@ -1,5 +1,5 @@
 from typing import List
 
-__version__ = "1.2.1"
+__version__ = "1.3.0.rc1"
 
 __all__: List[str] = []
