+++ joss/paper.md
@@ -3,17 +3,18 @@ title: 'Learning from Crowds with Crowd-Kit'
 tags:
   - Python
   - crowdsourcing
+  - data labeling
+  - answer aggregation
   - truth inference
-  - aggregation
   - learning from crowds
   - machine learning
   - quality control
+  - data quality
 authors:
   - name: Dmitry Ustalov
     orcid: 0000-0002-9979-2188
     affiliation: 1
     corresponding: true
-    note: The work was done while Dmitry and Boris were with Toloka.
   - name: Nikita Pavlichenko
     orcid: 0000-0002-7330-393X
     affiliation: 2
@@ -23,7 +24,7 @@ authors:
 affiliations:
  - name: JetBrains, Serbia
    index: 1
- - name: Toloka, Serbia
+ - name: JetBrains, Germany
    index: 2
  - name: Planet Farms, Portugal
    index: 3
@@ -33,15 +34,19 @@ bibliography: paper.bib
 
 # Summary
 
-Quality control is a crux of crowdsourcing. While most means for quality control are organizational and imply worker selection, golden tasks, and post-acceptance, computational quality control techniques allow parameterizing the whole crowdsourcing process of workers, tasks, and labels, inferring and revealing relationships between them. In this paper, we present Crowd-Kit, a general-purpose crowdsourcing computational quality control toolkit. It provides efficient implementations in Python of computational quality control algorithms for crowdsourcing, including data quality estimators and truth inference methods. We focus on aggregation methods for all the major annotation tasks, from the categorical annotation in which latent label assumption is met to more complex tasks like image and sequence aggregation. We perform an extensive evaluation of our toolkit on several datasets of different natures, enabling benchmarking computational quality control methods in a uniform, systematic, and reproducible way using the same codebase. We release our code and data under an open-source license at <https://github.com/Toloka/crowd-kit>.
+This paper presents Crowd-Kit, a general-purpose computational quality control toolkit for crowdsourcing. Crowd-Kit provides efficient and convenient implementations of popular quality control algorithms in Python, including methods for truth inference, deep learning from crowds, and data quality estimation. Our toolkit supports multiple modalities of answers and provides dataset loaders and example notebooks for faster prototyping. We extensively evaluated our toolkit on several datasets of different natures, enabling benchmarking computational quality control methods in a uniform, systematic, and reproducible way using the same codebase. We release our code and data under the Apache License 2.0 at <https://github.com/Toloka/crowd-kit>.
 
 # Statement of need
 
[…]
