+++ joss/paper.bib
@@ -447,6 +447,20 @@
   language  = {english},
 }
 
+@article{Zhdanovskaya:23,
+  author    = {Zhdanovskaya, Anastasia and Baidakova, Daria and Ustalov, Dmitry},
+  title     = {{Data Labeling for Machine Learning Engineers: Project-Based Curriculum and Data-Centric Competitions}},
+  year      = {2023},
+  journal   = {Proceedings of the AAAI Conference on Artificial Intelligence},
+  volume    = {37},
+  number    = {13},
+  pages     = {15886--15893},
+  publisher = {AAAI Press},
+  issn      = {2374-3468},
+  doi       = {10.1609/aaai.v37i13.26886},
+  language  = {english},
+}
+
 @article{Zheng:17,
   author    = {Zheng, Yudian and Li, Guoliang and Li, Yuanbing and Shan, Caihua and Cheng, Reynold},
   title     = {{Truth Inference in Crowdsourcing: Is the Problem Solved?}},
