+++ CHANGELOG.md
@@ -57,13 +57,11 @@
 * Added ROVER
 * Fixed HRRASA and refactored TextRASA and TextHRRASA
 
-
 0.0.6
 -----
 
 * Added realization of inter-annotator agreement coefficient (Krippendorff 1980): `alpha_krippendorff`
 * Added usage examples
-
 
 0.0.5
 -----
