+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include LICENSE AUTHORS CONTRIBUTING.md
+include LICENSE AUTHORS CONTRIBUTING.md CITATION.cff
