### 2024-04-05 d3eb4ff
Bump version to 1.3.0
