### 2024-04-05 #6227 @dustalov
@arfon: we read the paper, released the new version to PyPI, and uploaded the source archive to Zenodo.

- https://pypi.org/project/crowd-kit/1.3.0/
- https://zenodo.org/records/10934189
