### 2024-04-06 #6227 @dustalov
We have updated the references and citations in our repositories. Also, I updated the arXiv preprint.

@arfon, @jorgedch, @Mind-the-Cap, @mitchellg: thank you very much! I wish more journals follow the JOSS model.
